Chore(theme): move Colors file to theme folder
Add global styles for reusable classes
