+++ src/config/theme/globalStyles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+import { Colors } from "./Colors";
+
+export const globalStyles = StyleSheet.create({
+    title: {
+        fontSize: 35,
+        fontWeight: 'bold',
+        textAlign: 'center',
+        marginBottom: 20,
+        marginVertical: 5,
+        color: Colors.text
+    },
+});
