+++ src/presentation/navigation/StackNavigator.tsx
@@ -1,6 +1,6 @@
 import { CardStyleInterpolators, createStackNavigator } from '@react-navigation/stack';
 import { CartListScreen } from '../screens/cart/CartListScreen';
-import { Colors } from '../../config/Colors';
+import { Colors } from '../../config/theme/Colors';
 import { NewCartScreen } from '../screens/cart/NewCartScreen';
 import { StackNavbar } from '../components/StackNavbar';
 import { CartDetailsScreen } from '../screens/cart/CartDetailsScreen';
